// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/htmltemplate-type-model.ts
export class HtmlTemplateModel {
    isArchived: any;
    timeStamp: any;
    htmlTemplateId: string;
    htmlTemplateName: string;
    htmlTemplate: string;
    isConfigurable: boolean;
    isMaster: boolean;
    isRoleBased: boolean;
    isMailBased: boolean;
    selectedRoleIds: string[];
    selectedRoleNames: string;
    mailUrls: string[];
    mails: string[];
}